scan_ip/requirements.txt
psycopg2-binary
python-nmap
kafka-python
python-dotenv
schedule
